+++ src/os/mod.rs
@@ -1,3 +1,6 @@
+use std::error::Error;
+use std::fmt;
+
 #[cfg(target_os = "macos")]
 mod macos;
 #[cfg(target_os = "macos")]
@@ -6,3 +9,14 @@ pub(crate) use macos::*;
 mod unix;
 #[cfg(all(unix, not(target_os = "macos")))]
 pub(crate) use unix::*;
+
+#[derive(Debug)]
+struct PollReadTimedOutError;
+
+impl fmt::Display for PollReadTimedOutError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "poll_read timed out")
+    }
+}
+
+impl Error for PollReadTimedOutError {}