+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 ## 0.4.2
 * Add `keywords` to package metadata.
+* Remove dependency on `thiserror`.
 
 ## 0.4.1
 * 🐛 Fixed `OSC 11` response being visible to users of GNU Screen